Re-add extension/*Makefile changes to put this PR in final form
The previous run worked
https://github.com/timescale/timescale-prometheus/pull/78/checks?check_run_id=650529500

The next run will show us if we're double-running the tests, but that
may be the best we can do.

We deliberately test with -v when docker changed, because if there's
and error in extension-creation, we want to see what it is.

We should add a push to master with extension changed action that pushes
the new docker image